isolate osql prefetch in its own unit
